Renamed plugin.
