+++ pom.xml
@@ -8,7 +8,7 @@
         <relativePath />
     </parent>
     <groupId>io.jenkins.plugins</groupId>
-    <artifactId>pipeline-log-fluentd-cloudwatch</artifactId> <!-- TODO minus fluentd -->
+    <artifactId>pipeline-cloudwatch-logs</artifactId>
     <version>${revision}${changelist}</version>
     <packaging>hpi</packaging>
     <properties>
@@ -27,7 +27,7 @@
             <url>https://opensource.org/licenses/MIT</url>
         </license>
     </licenses>
-    <url>https://wiki.jenkins.io/display/JENKINS/Pipeline+Log+Fluentd+CloudWatch+Plugin</url> <!-- TODO minus fluentd -->
+    <url>https://wiki.jenkins.io/display/JENKINS/Pipeline+CloudWatch+Logs+Plugin</url>
     <scm>
         <connection>scm:git:git://github.com/jenkinsci/${project.artifactId}-plugin.git</connection>
         <developerConnection>scm:git:git@github.com:jenkinsci/${project.artifactId}-plugin.git</developerConnection>
@@ -180,7 +180,7 @@
                 <artifactId>maven-hpi-plugin</artifactId>
                 <configuration>
                     <loggers>
-                        <io.jenkins.plugins.pipeline_log_fluentd_cloudwatch>FINE</io.jenkins.plugins.pipeline_log_fluentd_cloudwatch>
+                        <io.jenkins.plugins.pipeline_cloudwatch_logs>FINE</io.jenkins.plugins.pipeline_cloudwatch_logs>
                     </loggers>
                 </configuration>
             </plugin>
