+++ src/main/java/io/jenkins/plugins/pipeline_cloudwatch_logs/CloudWatchSender.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package io.jenkins.plugins.pipeline_log_fluentd_cloudwatch;
+package io.jenkins.plugins.pipeline_cloudwatch_logs;
 
 import com.amazonaws.services.logs.model.InputLogEvent;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
