+++ src/main/java/io/jenkins/plugins/pipeline_cloudwatch_logs/ConsoleNotes.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package io.jenkins.plugins.pipeline_log_fluentd_cloudwatch;
+package io.jenkins.plugins.pipeline_cloudwatch_logs;
 
 import com.google.common.collect.ImmutableMap;
 import hudson.console.ConsoleNote;
