+++ src/main/java/io/jenkins/plugins/pipeline_cloudwatch_logs/PipelineBridge.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package io.jenkins.plugins.pipeline_log_fluentd_cloudwatch; // TODO minus fluentd
+package io.jenkins.plugins.pipeline_cloudwatch_logs;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import hudson.Extension;
