+++ src/test/java/io/jenkins/plugins/pipeline_cloudwatch_logs/CloudWatchAwsGlobalConfigurationStub.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License
  *
- * Copyright 2018 CloudBees, Inc.
+ * Copyright (c) 2018, CloudBees, Inc.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -22,20 +22,17 @@
  * THE SOFTWARE.
  */
 
-package io.jenkins.plugins.pipeline_log_fluentd_cloudwatch;
+package io.jenkins.plugins.pipeline_cloudwatch_logs;
 
-import org.junit.Test;
-import static org.junit.Assert.*;
+import com.amazonaws.services.logs.AWSLogs;
 
-public class TimestampTrackerTest {
+public class CloudWatchAwsGlobalConfigurationStub extends CloudWatchAwsGlobalConfiguration {
 
-    @Test public void monoticallyIncrease() throws Exception {
-        assertEquals(1234, TimestampTracker.monoticallyIncrease(0, 1234));
-        assertEquals(1235, TimestampTracker.monoticallyIncrease(1234, 1234));
-        assertEquals(1236, TimestampTracker.monoticallyIncrease(1235, 1234));
-        assertEquals(3000, TimestampTracker.monoticallyIncrease(3000, 1234));
+    public CloudWatchAwsGlobalConfigurationStub() {
+        super(true);
     }
 
-    // TODO test for checkCompletion
-
+    protected void filter(AWSLogs client, String logGroupName2) {
+        // noop
+    }
 }
